+++ src/main/java/com/minhcraft/ClassicReintegratedTweaks.java
@@ -2,7 +2,6 @@ package com.minhcraft;
 
 import com.minhcraft.config.ModConfig;
 import com.minhcraft.config.RecipeCustomSortingConfigLoader;
-import com.minhcraft.register.ModEvents;
 import com.minhcraft.register.ModItems;
 import com.minhcraft.register.ModRegistry;
 import com.minhcraft.register.ModSounds;
@@ -30,7 +29,6 @@ public class ClassicReintegratedTweaks implements ModInitializer {
 
 		LOGGER.info("Initializing Classic Reintegrated Tweaks");
 		ModSounds.init();
-		ModEvents.init();
 		ModItems.init();
 		ModRegistry.init();
 		RecipeCustomSortingConfigLoader.loadConfig();