+++ src/main/java/com/minhcraft/mixin/entity/SpawnPlacementsMixin.java
@@ -1,0 +1,25 @@
+package com.minhcraft.mixin.entity;
+
+import com.minhcraft.entity.GiantSpawnRules;
+import net.minecraft.world.entity.SpawnPlacements;
+import net.minecraft.world.entity.monster.Monster;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.*;
+
+@Mixin(SpawnPlacements.class)
+public abstract class SpawnPlacementsMixin {
+
+    // Modify giant spawn rules
+    @ModifyArg(
+            method = "<clinit>",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/SpawnPlacements;register(Lnet/minecraft/world/entity/EntityType;Lnet/minecraft/world/entity/SpawnPlacements$Type;Lnet/minecraft/world/level/levelgen/Heightmap$Types;Lnet/minecraft/world/entity/SpawnPlacements$SpawnPredicate;)V"),
+            slice = @Slice(
+                    from = @At(value = "FIELD", target = "Lnet/minecraft/world/entity/EntityType;GIANT:Lnet/minecraft/world/entity/EntityType;"),
+                    to = @At(value = "FIELD", target = "Lnet/minecraft/world/entity/EntityType;GLOW_SQUID:Lnet/minecraft/world/entity/EntityType;")
+            ),
+            index = 3
+    )
+    private static SpawnPlacements.SpawnPredicate<? extends Monster> test(SpawnPlacements.SpawnPredicate<? extends Monster> decoratorPredicate) {
+        return GiantSpawnRules::checkGiantSpawnRules;
+    }
+}